# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "inventory-api",
    "description": "A minimal api for inventory management",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/": {
      "get": {
        "description": "Displays the home page",
        "parameters": [],
        "responses": {}
      }
    },
    "/items/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "Creates a new item in the DB",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": { "$ref": "#/components/schemas/items" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/items/{itemId}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": { "$ref": "#/components/schemas/items" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/warehouses/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": { "$ref": "#/components/schemas/warehouses" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/warehouses/{warehouseId}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "warehouseId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": { "$ref": "#/components/schemas/warehouses" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "warehouseId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "items": {
        "type": "object",
        "properties": {
          "WarehouseId": {
            "example": 1,
            "type": "number"
          },
          "name": {
            "example": "any"
          },
          "quantity": {
            "example": 8000,
            "type": "number"
          },
          "price": {
            "example": 19.99,
            "type": "number"
          }
        }
      },
      "warehouses": {
        "type": "object",
        "properties": {
          "name": {
            "example": "Amazon2389",
            "type": "string"
          },
          "city": {
            "example": "Springfield",
            "type": "string"
          },
          "state": {
            "example": "Illinois",
            "type": "string"
          }
        }
      }
    }
  }
}